loan edit issue

+++ app/Services/Admin/LoanService.php
@@ -463,7 +463,7 @@ class LoanService
         $interestPerMonth   = round($principal * $monthlyRate, 2);
         $initiationPerMonth = round($totalInitiation / $term, 2);
 
-        $payday = $loan->salary_payday ?? 25;
+        $payday = (int) ($loan->salary_payday ?? 25);
         $payDate = \Carbon\Carbon::parse($loan->disbursement_date)->addMonth()->setDay($payday);
 
         for ($i = 1; $i <= $term; $i++) {
